Add create e-mail notificator modal box in Gamecraft UI

+++ gamecraft-ui/src/main/resources/static/js/ui_operations.js
@@ -165,14 +165,26 @@ function fillEmailAccountsTable() {
     var accounts = getEmailAccounts();
     $(function() {
         $.each(accounts, function(i, item) {
-            var tr = $('<tr>').append(
-                $('<tr>'),
-                $('<td>').text(item.id),
-                $('<td>').text(item.emailAccountName),
-                $('<td>').text(item.emailAccountDescription),
-                $('<td>').html("<button type=\"button\" class=\"btn btn-danger btn-xs\" onclick=\"deleteEmailAccount('" +item.id + "'); location.reload(); \"><span class=\"glyphicon glyphicon-remove\"></span> </button>")
-            );
-            $(".table").append(tr.html());
+            if (isAdmin(getUsername())) {
+                var tr = $('<tr>').append(
+                    $('<tr>'),
+                    $('<td>').text(item.id),
+                    $('<td>').text(item.emailAccountName),
+                    $('<td>').text(item.emailAccountDescription),
+                    $('<td>').html("<button type=\"button\" class=\"btn btn-danger btn-xs\" onclick=\"deleteEmailAccount('" + item.id + "'); location.reload(); \"><span class=\"glyphicon glyphicon-remove\"></span> </button>")
+                );
+                $(".table").append(tr.html());
+            }
+            else {
+                var tr = $('<tr>').append(
+                    $('<tr>'),
+                    $('<td>').text(item.id),
+                    $('<td>').text(item.emailAccountName),
+                    $('<td>').text(item.emailAccountDescription),
+                    $('<td>').html("")
+                );
+                $(".table").append(tr.html());
+            }
         });
     });
 }
